#6 土台の作成
